+++ assets/javascripts/kitten/karl/organisms/loan-simulator.js
@@ -52,6 +52,25 @@ const KarlLoanSimulatorWithCommission = defaultProps(KarlLoanSimulator, {
   actionLabel: null,
 })
 
+const KarlLoanSimulatorWithFeesExemptions = defaultProps(KarlLoanSimulator, {
+  displayCommission: true,
+  feesExemptions: true,
+  feesExemptionsLabel: `au lieu de `,
+  commissionLabel: `Commission${nbsp}:`,
+
+  commissionRules: [
+    { durationMax: 9, rate: 0.03 },
+    { durationMax: 18, rate: 0.04 },
+    { durationMax: 24, rate: 0.05 },
+    { rate: 0.06 },
+  ],
+
+  initialAmount: 2500,
+  initialInstallment: 150,
+
+  actionLabel: null,
+})
+
 const KarlLoanSimulatorWithError = defaultProps(KarlLoanSimulator, {
   initialTouched: true,
   actionLabel: null,
@@ -59,4 +78,5 @@ const KarlLoanSimulatorWithError = defaultProps(KarlLoanSimulator, {
 
 export { KarlLoanSimulator,
          KarlLoanSimulatorWithCommission,
+         KarlLoanSimulatorWithFeesExemptions,
          KarlLoanSimulatorWithError }