+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Features: Add exemption to `LoanSimulator`.
+
 ## [10.1.0] - 2017-04-11
 
 Fix: